file renaming + package rearranging
