+++ src/main/java/cz/daiton/foodsquare/exceptions/CustomFieldError.java
@@ -1,4 +1,4 @@
-package cz.daiton.foodsquare.security;
+package cz.daiton.foodsquare.exceptions;
 
 public class CustomFieldError {
