+++ src/main/java/cz/daiton/foodsquare/like/LikeServiceImpl.java
@@ -5,7 +5,7 @@ import cz.daiton.foodsquare.appuser.AppUserRepository;
 import cz.daiton.foodsquare.appuser.AppUserService;
 import cz.daiton.foodsquare.post.Post;
 import cz.daiton.foodsquare.post.PostRepository;
-import cz.daiton.foodsquare.security.IncorrectUserException;
+import cz.daiton.foodsquare.exceptions.IncorrectUserException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
