+++ src/main/java/cz/daiton/foodsquare/post/PostService.java
@@ -1,6 +1,6 @@
 package cz.daiton.foodsquare.post;
 
-import cz.daiton.foodsquare.security.IncorrectUserException;
+import cz.daiton.foodsquare.exceptions.IncorrectUserException;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
