+++ src/main/java/cz/daiton/foodsquare/post/review/ReviewController.java
@@ -3,7 +3,7 @@ package cz.daiton.foodsquare.post.review;
 import cz.daiton.foodsquare.appuser.AppUserService;
 import cz.daiton.foodsquare.payload.response.MessageResponse;
 import cz.daiton.foodsquare.payload.response.PostContentResponse;
-import cz.daiton.foodsquare.security.IncorrectUserException;
+import cz.daiton.foodsquare.exceptions.IncorrectUserException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.security.access.prepost.PreAuthorize;
