+++ src/main/java/cz/daiton/foodsquare/security/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package cz.daiton.foodsquare.security.userdetails;
+package cz.daiton.foodsquare.security;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import cz.daiton.foodsquare.appuser.AppUser;
